+++ src/project2/ver04/NormalAccount.java
@@ -2,7 +2,7 @@ package project2.ver04;
 
 public class NormalAccount extends Account {
 
-	public NormalAccount(String number, String name, double balance, int rate) {
+	public NormalAccount(String number, String name, int balance, int rate) {
 		super(number, name, balance);
 
 		this.rate = rate;
@@ -12,7 +12,7 @@ public class NormalAccount extends Account {
 		System.out.println("-------");
 		System.out.println("계좌번호>" + accountNumber);
 		System.out.println("고객이름>" + name);
-		System.out.println("잔고>" + allMoney);
+		System.out.println("잔고>" + (int)allMoney);
 		System.out.println("기본이자>" + rate + "%");
 		System.out.println("-------");
 	}
